createGenerateOptions {CirceR}	R Documentation

Create Generation Options

Description

Creates the generation options object for use in generateSql()

Usage

createGenerateOptions(
  cohortIdFieldName,
  cohortId,
  cdmSchema,
  targetTable,
  resultSchema,
  vocabularySchema,
  generateStats
)

Arguments

cohortIdFieldName
The field that contains the cohortId in the cohort table.

cohortId
The generated cohort ID.

cdmSchema
The value of the CDM schema.

targetTable
the cohort table name.

resultSchema
the schema the cohort table belongs to.

vocabularySchema
the schema of the vocabulary tables (defaults to cdmSchema)

generateStats
a boolean representing if the query should include inclusion rule statistics calculation

Value

A jObj that can be passed to generateSql().

[Package CirceR version 1.1.1 Index]